requirements.txt
pybrain
nose
nltk
scipy
python-dateutil
